Grow Article on Hover. Action Button Event listeners moved to buttons.py
