+++ index.html
@@ -52,8 +52,8 @@
 
         <main></main>
         <footer>
-            <button py-click="see_less_button_clicked">&lt See Less &gt</button>
-            <button py-click="see_more_button_clicked">&lt See More &gt</button>
+            <button class="load-button" py-click="see_less_button_clicked">&lt See Less &gt</button>
+            <button class="load-button" py-click="see_more_button_clicked">&lt See More &gt</button>
         </footer>
         <script type="py" src="./index.py" config="./pyscript.toml"></script>
     </body>
